+++ ci/bin/deploy_control_lint.dart
@@ -1,0 +1,8 @@
+import 'package:ci/ci.dart' as ci;
+
+void main(List<String> arguments) async {
+  await ci.runAsync(
+    'deploy control_lint',
+    () => ci.deployDart('control_lint'),
+  );
+}